requirements.txt
Flask==3.0.3
strawberry-graphql[debug-server]